Update UANodeSetHelpers.cs
In OPC UA V1.04, there were new access level bits defined. In the OPC UA information model, the attribute AccessLevel (byte) was not changed. Instead, a new optional attribute AccessLevelEx (Unsigned32) was added, whose lower 8 bits are identical to the the attribute AccessLevel.

In the XML NodeSet file, a different approach was chosen. Instead of adding another attribute, the existing attribute AccessLevel was changed from Byte to Unsigned32.
See https://opcfoundation.org/UA/schemas/1.04/UANodeSet.xsd.

+++ Stack/Opc.Ua.Core/Schema/UANodeSetHelpers.cs
@@ -472,7 +472,7 @@ namespace Opc.Ua.Export
                     value.DataType = ImportNodeId(o.DataType, context.NamespaceUris, true);
                     value.ValueRank = o.ValueRank;
                     value.ArrayDimensions = ImportArrayDimensions(o.ArrayDimensions);
-                    value.AccessLevel = o.AccessLevel;
+                    value.AccessLevel = (byte)o.AccessLevel;
                     value.UserAccessLevel = o.UserAccessLevel;
                     //if UserAccessLevel is not specified but Access level is diferent from default use AccessLevel for UserAccessLevel
                     if (o.UserAccessLevel == AccessLevels.CurrentRead && o.AccessLevel > o.UserAccessLevel)
